requirements.txt
Flask==2.3.3
requests==2.31.0
flask-cors==3.0.10
python-dotenv==1.0.0
gunicorn==20.1.0
